// payload-types.ts
export interface Media {
  id: string;
  alt: string;
  credit?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    card?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    hero?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: string;
  name: string;
  /**
   * Used in the URL. Leave blank to generate from the title.
   */
  slug: string;
  description: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "authors".
 */
export interface Author {
  id: string;
  name: string;
  /**
   * Used in the URL. Leave blank to generate from the title.
   */
  slug: string;
  role: string;
  /**
   * Initials. Auto-filled from the name if left blank.
   */
  avatar?: string | null;
  photo?: (string | null) | Media;
  bio: string;
  twitter?: string | null;
  linkedin?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "articles".
 */
export interface Article {
  id: string;
  title: string;
  /**
   * Used in the URL. Leave blank to generate from the title.
   */
  slug: string;
  deck: string;
  image: string | Media;
  content: (
    | {
        value: string;
        id?: string | null;
        blockName?: string | null;
        blockType: 'paragraph';
      }
    | {
        value: string;
        id?: string | null;
        blockName?: string | null;
        blockType: 'subheading';
      }
    | {
        value: string;
        attribution?: string | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'pullquote';
      }
    | {
        image: string | Media;
        caption?: string | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'image';
      }
    | {
        value: string;
        caption?: string | null;
        id?: string | null;
        blockName?: string | null;
        blockType: 'embed';
      }
  )[];
  category: string | Category;
  author: string | Author;
  publishedAt: string;
  readTime?: string | null;
  views?: number | null;
  featured?: boolean | null;
  trending?: boolean | null;
  editorsPick?: boolean | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}